pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for ikatoo.com.br"
authors = ["Milton Carlos Katoo <mckatoo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
flake8 = "^7.0.0"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
django-types = "^0.19.1"
django-summernote = "^0.8.20.0"
djangorestframework = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
